+++ components/MediaUpload/MediaUpload.tsx
@@ -1,10 +1,8 @@
 import ImageUpload from "./ImageUpload";
-import AudioUpload from "./AudioUpload";
 
 const MediaUpload = (props:any) =>  (
   <>
-    <ImageUpload label="Upload NFT Art (Required)" nftImage={props.nftImage} setNftImage={props.setNftImage} formRegisterName="nftImage" />
-    <AudioUpload audioFile={props.audioFile} setAudioFile={props.setAudioFile} formRegisterName="audioFile" />
+  <ImageUpload label="Upload NFT Art (Required)" nftImage={props.nftImage} setNftImage={props.setNftImage} formRegisterName="nftImage" />
   </>
 )
