+++ components/Navbar/navbar.module.css
@@ -1,0 +1,33 @@
+.navbar {
+  z-index: 1000;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  display: flex;
+  align-items: center;
+  padding: 12px 24px;
+  gap: 10px;
+  min-height: 64px;
+  background-color: rgba(255, 255, 255, .1);
+  backdrop-filter: blur(8px);
+  justify-content: space-between;
+  border-bottom: 1px solid rgb(229, 231, 235);
+}
+
+.txScreen ~ .navbar {
+  background-color: rgba(0, 0, 0, .1);
+  color: white;
+  stroke: white;
+  fill: white;
+}
+
+.navbar.darkMode {
+  background-color: rgba(0, 0, 0, .1);
+}
+
+.darkMode {
+  color: white;
+  stroke: white;
+  fill: white;
+}
