+++ env.local
@@ -1,0 +1,6 @@
+NEXTAUTH_SECRET= # Mac/Linux: `openssl rand -hex 32` or go to https://generate-secret.now.sh/32
+
+# development
+NEXTAUTH_URL=http://localhost:3000
+
+NEXT_PUBLIC_ENABLE_TESTNETS=true
