+++ pages/_app.tsx
@@ -5,6 +5,8 @@ import { RainbowKitProvider, getDefaultWallets } from '@rainbow-me/rainbowkit';
 import { chain, configureChains, createClient, WagmiConfig } from 'wagmi';
 import { alchemyProvider } from 'wagmi/providers/alchemy';
 import { publicProvider } from 'wagmi/providers/public';
+import Navbar from '../components/Navbar/Navbar';
+
 
 const { chains, provider, webSocketProvider } = configureChains(
   [
@@ -26,7 +28,7 @@ const { chains, provider, webSocketProvider } = configureChains(
 );
 
 const { connectors } = getDefaultWallets({
-  appName: 'Start Decent',
+  appName: 'AI NFTs',
   chains,
 });
 
@@ -41,6 +43,7 @@ function MyApp({ Component, pageProps }: AppProps) {
   return (
     <WagmiConfig client={wagmiClient}>
       <RainbowKitProvider chains={chains}>
+        <Navbar darkMode />
         <Component {...pageProps} />
       </RainbowKitProvider>
     </WagmiConfig>
